requirements.txt
python=3.7
wandb
logzero
pytorch
torchvision
jupyter
albumentations
opencv
matplotlib
pytorch-cuda
numpy
pandas
timm
scikit-learn
tqdm
cachetools
